requirements.txt
Pillow==9.0.0
plotly==5.5.0
kaleido==0.2.1
